Added toolbar button, and refactor
