--- lib/src/main/java/com/nikialeksey/fullscreendialog/Button.java
@@ -1,5 +1,0 @@
-package com.nikialeksey.fullscreendialog;
-
-public interface Button {
-    void addOnClick(ClickListener clickListener);
-}
