+++ lib/src/main/java/com/nikialeksey/fullscreendialog/Dialog.java
@@ -1,0 +1,10 @@
+package com.nikialeksey.fullscreendialog;
+
+import android.content.DialogInterface;
+import android.support.annotation.NonNull;
+
+public interface Dialog extends DialogInterface {
+    void show();
+
+    void addOnClose(@NonNull ClickListener click);
+}
