+++ lib/src/main/java/com/nikialeksey/fullscreendialog/FsDialogToolbar.java
@@ -1,20 +1,12 @@
 package com.nikialeksey.fullscreendialog;
 
-
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.graphics.PorterDuff;
-import android.graphics.drawable.Drawable;
 import android.support.annotation.NonNull;
 import android.support.design.widget.AppBarLayout;
-import android.support.v7.content.res.AppCompatResources;
 import android.support.v7.widget.Toolbar;
-import android.view.MenuItem;
-import android.view.View;
-
+import com.nikialeksey.fullscreendialog.buttons.ToolbarButton;
 import com.nikialeksey.fullscreendialog.theme.Color;
-import java.util.ArrayList;
-import java.util.List;
 
 import static android.view.ViewGroup.LayoutParams.MATCH_PARENT;
 import static android.view.ViewGroup.LayoutParams.WRAP_CONTENT;
@@ -22,13 +14,13 @@ import static android.view.ViewGroup.LayoutParams.WRAP_CONTENT;
 @SuppressLint("ViewConstructor")
 public class FsDialogToolbar extends AppBarLayout {
 
-    private final FsCloseButton fsCloseButton;
+    private final ToolbarButton fsCloseButton;
 
     public FsDialogToolbar(@NonNull final Context context, @NonNull final String title,
-                           @NonNull final FsCloseButton fsCloseButton,
-                           @NonNull final FsActionButton fsActionButton) {
+                           @NonNull final ToolbarButton closeButton,
+                           @NonNull final ToolbarButton actionButton) {
         super(context);
-        this.fsCloseButton = fsCloseButton;
+        this.fsCloseButton = closeButton;
 
         final Color textColorPrimary = new Color(context, android.R.attr.textColorPrimary);
 
@@ -36,13 +28,13 @@ public class FsDialogToolbar extends AppBarLayout {
         toolbar.setTitle(title);
         toolbar.setTitleTextColor(textColorPrimary.intValue());
 
-        fsCloseButton.addInToolbar(toolbar);
-        fsActionButton.addInToolbar(toolbar);
+        closeButton.addInToolbar(toolbar);
+        actionButton.addInToolbar(toolbar);
 
         addView(toolbar, new AppBarLayout.LayoutParams(MATCH_PARENT, WRAP_CONTENT));
     }
 
-    public void addOnClose(final ClickListener clickListener) {
+    public void addOnClose(@NonNull final ClickListener clickListener) {
         fsCloseButton.addOnClick(clickListener);
     }
 }
