+++ lib/src/main/java/com/nikialeksey/fullscreendialog/buttons/Button.java
@@ -1,0 +1,10 @@
+package com.nikialeksey.fullscreendialog.buttons;
+
+import android.support.annotation.NonNull;
+import com.nikialeksey.fullscreendialog.ClickListener;
+
+public interface Button {
+    void addOnClick(@NonNull ClickListener clickListener);
+
+    void click();
+}
