+++ lib/src/main/java/com/nikialeksey/fullscreendialog/buttons/ToolbarButton.java
@@ -1,0 +1,8 @@
+package com.nikialeksey.fullscreendialog.buttons;
+
+import android.support.annotation.NonNull;
+import android.support.v7.widget.Toolbar;
+
+public interface ToolbarButton extends Button {
+    void addInToolbar(@NonNull Toolbar toolbar);
+}
